Static compilation before ONNX codegen
Run the static Ten compiler and type checked before compiling to ONNX output.
