+++ test/test_onnx.py
@@ -1,34 +1,32 @@
 # Copyright 2023 Luke Hoban
 
 import unittest
-from ten import parse, onnx_wip
+from ten import parse, onnx_wip, compiler, builtins
 import numpy as np
 import onnxruntime as ort
 import onnx.printer as onnx_printer
 import onnx.shape_inference as onnx_shape_inference
 import baseline
+from typing import Mapping
 
 
 class OnnxCompilerTestCase(unittest.TestCase):
     def test_linear(self):
-        p = parse.Parser(
-            """
+        p = """
             Linear[N,K]|w:{N,K},b:{K}|(x:{...,N}) -> {...,K}:
                 return x@w + b
             """
-        )
-        decls = p.parse_program()
-        linear_decl = decls[0]
-        compiler = onnx_wip.Compiler()
+
         static_args = {"N": 10, "K": 7}
         params = {"w": np.ones([10, 7]), "b": np.ones([7])}
-        model = compiler.compile_program(decls, linear_decl, static_args, params)
-        print(onnx_printer.to_text(model))
-        ort_sess = ort.InferenceSession(model.SerializeToString())
-        outputs = ort_sess.run(
-            None,
+
+        outputs = self.compile_and_run(
+            p,
+            static_args,
+            params,
             {"x": [[1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]]},
         )
+
         expected = [
             [56.0, 56.0, 56.0, 56.0, 56.0, 56.0, 56.0],
             [56.0, 56.0, 56.0, 56.0, 56.0, 56.0, 56.0],
@@ -36,24 +34,17 @@ class OnnxCompilerTestCase(unittest.TestCase):
         np.testing.assert_array_almost_equal(outputs[0], np.array(expected))
 
     def test_gelu(self):
-        p = parse.Parser(
-            """
+        p = """
             Gelu(x: {...}) -> {...}:
                 return 0.5 * x * (1 + Tanh(0.7978845608 * x + 0.044715 * x**3))
             """
-        )
-        compiler = onnx_wip.Compiler()
-        decls = p.parse_program()
-        gelu_decl = decls[0]
+
         static_args = {}
         params = {}
-        model = compiler.compile_program(decls, gelu_decl, static_args, params)
-        print(onnx_printer.to_text(model))
-        ort_sess = ort.InferenceSession(model.SerializeToString())
+
         for x, y in [(-1, -0.156408), (0.0, 0.0), (1.0, 0.843592), (2.0, 1.96059)]:
-            outputs = ort_sess.run(
-                None,
-                {"x": np.full([], x).astype(np.float32)},
+            outputs = self.compile_and_run(
+                p, static_args, params, {"x": np.full([], x).astype(np.float32)}
             )
             np.testing.assert_array_almost_equal(
                 outputs[0],
@@ -71,7 +62,7 @@ class OnnxCompilerTestCase(unittest.TestCase):
             FFN[S,E]|c_fc, c_proj|(x:{S,E}) -> {S,E}:
                 return Linear[E*4,E]|c_proj|(Gelu(Linear[E,E*4]|c_fc|(x)))
             """
-        static_args = {"S": 3, "E": 2}
+        static_args = {"S": 3.0, "E": 2.0}
         params = {
             "c_fc": {"w": np.ones([2, 8]), "b": np.ones([8])},
             "c_proj": {"w": np.ones([8, 2]), "b": np.ones([2])},
@@ -109,7 +100,7 @@ class OnnxCompilerTestCase(unittest.TestCase):
                 out = Attention[S,K,S,K](q, k, v, causal_mask) {H,S,K -> S,(H,K)}
                 return Linear[E,E]|c_proj|(out)
             """
-        static_args = {"H": 2, "S": 3, "E": 4, "K": 2}
+        static_args = {"H": 2.0, "S": 3.0, "E": 4.0, "K": 2.0}
         params = {
             "c_attn": {"w": np.ones([4, 12]), "b": np.ones([12])},
             "c_proj": {"w": np.ones([4, 4]), "b": np.ones([4])},
@@ -160,7 +151,7 @@ class OnnxCompilerTestCase(unittest.TestCase):
                 return y + FFN[S,E]|mlp|(LayerNorm[S,E]|ln_2|(y))
             """
         H, S, E = 2, 3, 4
-        static_args = {"H": H, "S": S, "E": E}
+        static_args = {"H": float(H), "S": float(S), "E": float(E)}
         params = {
             "mlp": {
                 "c_fc": {"w": np.ones([E, 4 * E]), "b": np.ones([4 * E])},
@@ -310,7 +301,7 @@ class OnnxCompilerTestCase(unittest.TestCase):
         R[S]() -> {S}:
             return Range[S]()
         """
-        static_args = {"S": 2}
+        static_args = {"S": 2.0}
         params = {}
         inputs = {}
         outputs = self.compile_and_run(p, static_args, params, inputs)
@@ -320,13 +311,26 @@ class OnnxCompilerTestCase(unittest.TestCase):
         np.testing.assert_array_almost_equal(outputs[0], expected, decimal=3)
 
     def compile_and_run(
-        self, program: str, static_args: dict, params: dict, inputs: dict
+        self, program: str, static_args: Mapping[str, float], params: dict, inputs: dict
     ):
         p = parse.Parser(program)
         decls = p.parse_program()
         decl = decls[-1]
-        compiler = onnx_wip.Compiler()
-        model = compiler.compile_program(decls, decl, static_args, params)
+
+        tencompiler = compiler.Compiler()
+        funcs = {func.name.text: func for func in decls}
+        type_env = compiler.TypeEnv(None, {}, {}, {**builtins.built_ins, **funcs})
+        compiled_decl = tencompiler.compile_function(
+            decl, list(static_args.values()), type_env
+        )
+        print(compiled_decl)
+        for f, func in tencompiler.funcs.items():
+            print(f"{f}: {func}")
+
+        onnxcompiler = onnx_wip.Compiler()
+        model = onnxcompiler.compile_program(
+            tencompiler.funcs, compiled_decl, static_args, params
+        )
         model = onnx_shape_inference.infer_shapes(
             model, check_type=True, data_prop=True
         )
